+++ modules/jarvos-secondbrain/adapters/hermes/index.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const {
+  createSessionSourceAdapter,
+} = require('../session-source/session-source-adapter');
+
+function createHermesSessionAdapter(options = {}) {
+  return createSessionSourceAdapter('hermes', options);
+}
+
+module.exports = {
+  createHermesSessionAdapter,
+};